refactor(explorer): share the dirty pivot configuration hook
The chart card and the embedded builder each rebuilt the dirty pivot
from the store, merge state and explore fields; one drifting copy
would split what "stale results" means between the two warnings.
